# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dunda club operating system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List active orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, PENDING, ACCEPTED, PREPARING, READY, SERVED, COMPLETED, CANCELLED]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
...
  schemas:
...
    Order:
      type: object
      required: [id, number, table, status, items, createdAt]
      properties:
        id: { type: string }
        number: { type: string }
        table: { type: ["string", "null"] }
        status: { type: string, enum: [DRAFT, PENDING, ACCEPTED, PREPARING, READY, SERVED, COMPLETED, CANCELLED] }
        items: { type: array, items: { type: string } }
        subtotal: { type: ["number", "null"] }
        serviceCharge: { type: ["number", "null"] }
        tax: { type: ["number", "null"] }
        discount: { type: ["number", "null"] }
        total: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CreateOrderInput:
      type: object
      required: [table, items]
      properties:
        table: { type: string }
        customerId: { type: ["string", "null"] }
        staffId: { type: ["string", "null"] }
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: string }
              quantity: { type: integer, minimum: 1 }
              unitId: { type: ["string", "null"] }
              notes: { type: ["string", "null"] }
    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [PENDING, ACCEPTED, PREPARING, READY, SERVED, COMPLETED, CANCELLED] }
...